feature: Change Readers interface to use Buffer protocol
This is in preparation to replacing Python implementation of protocol
parsing with a Rust implementation. To allow zero copy of bytes when
passing do from Python to Rust, we need to pass it as read-only objects.
By adhering to the Buffer protocol, we're also making sure to maximize
compatibility with various network layer implementations.

This is a breaking change, users of any reader functions must update
their code accordingly and now pass a Buffer + start offset instead of a
`io.IO[bytes]` object. Readers return the number of bytes they consume
off of the input.