Simplify serialization API
The serialization API is significicantly simplified. The original idea
of defining decoders as generators was a neat idea, and did have merit
in reducing duplication in that design.

However, there's need to question whether async decoders make sense.
Because in real cases, client code would multicast over a connection,
and would want to be able to free the connection as fast as possible, to
be able to read the next response. Because of that, it probably makes
more sense to read the full response into memory.