feature: Upgrade to build schema with 3.5.1
This commit upgrades to build schema from upstream tag 3.5.1.

The new version introduced two breakages that had to be fixed. One of
them is a model that is missing "versions" for a field, this was the
first such occasion in the entire schema, and was subsequently [fixed
upstream][fixed-upstream]. However, the discussion around the upstream
fix concludes there won't be any guarantees against identical breakage
in the future.

[fixed-upstream]: https://github.com/apache/kafka/pull/13680

The other breakage was a bit trickier to fix, and now requires to keep
track of future references in common structs during parsing. This was
solved with a recursive parser, looping until either all structs
successfully parse, or there is no more progress made.