Implement KIP-893 nullable entity fields (#101)
This commit implements support for the nullable entity field introduced
in `ConsumerGroupHeartbeatResponse.assignment`. This kind of field was
formalized upstream in KIP-893 but is not mentioned in any official
documentation of the protocol.

The KIP also mentions tagged fields. I left this out as I don't
understand why another mechanism for omiting a tagged field value is
needed, and AFAIK no entity uses this yet.

Partially addresses #100.