Fix behavior for default-only tagged entity fields (#102)
This enables the Java tests for the last non-record entity:
FetchRequest. This was previously failing because on the Java side, the
ReplicaState field tagged field is omitted when all of the nested fields
are equal to defaults.

Because our logic already has the correct behavior in general for tagged
fields, all that was needed here was to make such parent fields have a
default value. We detect when we can set a default by checking if the
nested entity has default values for all of its fields.

Partially addresses #100.