feature: Implement basic support for record fields (#122)
Issues to be discussed:

- This does not enable the Java compatibility tests for entities with
records.
These are currently failing as Hypothesis generates random bytes for the
records fields. We should most likely introduce a `RecordsBytes` type
and teach Hypothesis to generate some static record batch for this to
begin with (no need to have more intricate knowledge of record batches
yet).
- Java compatibility tests are failing when record batch fields are
`None`. Why?
- It's probably fine to merge the integration test with `@xfail`. It
needs to be solved at some point, but merging it doesn't make the
current situation any worse.