chore: Tune Hypothesis configuration
This addresses the issue that Java compatibility tests and other
roundtrip tests were configured with wildly different parameters. It
removes the explicit `@setting(max_examples=1)` from all round trip
tests, and now instead only configures this on a global level.

Adds two separate Hypothesis profiles and support for choosing them with
the `HYPOTHESIS_PROFILE` env var (I use [direnv] to set `exhaustive` in
my dev env).

[direnv]: https://direnv.net/

The `deterministic` is used by default, and it limits `max_examples` to
`5` which seems like a reasonable amount to start with. It enables
Hypothesis' own "deterministic" feature, which means it promises to
always generate the same data for a test function, as long as the test
function itself doesn't change. This is appropriate for CI.

Deadlines and the "too slow" healthcheck are disabled in both profiles,
as it's more likely to trigger on VM pauses than actual slow data
generation.