Only generate round-trippable timedelta values
This fixes a flakiness that was occurring occasionally in CI, with
timedelta values that caused roundtrip tests to fail.

Because timedelta is float based, we cannot guarantee arbitrary values
will survive a roundtrip through timedelta -> milliseconds -> timedelta.
We'll pragmatically ignore runtime effects of this for now (precision
loss is small anyway), but we need to take this into account to generate
test values with stable results.

The has_millisecond_precision check is removed from the timedelta type
predicates, it didn't really make sense there as it can elliminate valid
values that does actually roundtrip fine.