feature: Implement dynamic schema entity imports
This commit implements a set of helper functions to dynamically import
entities from the schema. They allow:

- Importing requests given their response type, and vice versa.
- Importing payload entities given their API key and type.
- Importing any schema entitiy given its schema name and type.

This is backed by an index that maps (schema name, version, entity type)
to an importable path name. The index is pre-generated together with the
rest of the schema, to limit impact on runtime code as much as possible.