Publish docs to Github Pages (#151)
This adds support for automatically publishing docs to Github Pages,
triggered by pushes to the main branch. This will only support a single
version of documentation, but we can look into extending support for
multiple versions in the future.

Thanks @TibsAtWork for pushing this! ⭐️

---

And! Forgot to mention, I've tested in this here:

- https://github.com/aiven-anton/kio (note the Deployments section, can
be clicked through)
- https://aiven-anton.github.io/kio/

Co-authored-by: Tibs <tibs@tonyibbs.co.uk>