chore: Remove version from schema docstrings (#164)
Please see separate commits, second one has the schema re-generation.

This addresses a regret. While it's nice to have the full link to the
upstream source easily available, this just causes an unjustified amount
of churn when rebuilding the schema for a new upstream version. I plan
to rebuild the schema for upstream version 3.7 soon, and I want to get
this merged beforehand. Being able to see exactly which APIs that
changes during a schema update will be more valuable then having the
convenience of the link.