fix: Always default tagged array field to the empty collection
Explanation: it seems we again differed with upstream in determining what
the "default" value of a complex field is. In this case, Apache Kafka
thinks a tagged field of a collection of nested entities, should
implicitly default to the empty collection. Previously this was guarded
by inspecting the "ignorable" value of fields, but this behavior was not
consistent with the new v10 Produce response.