refactor: Improve typing of classify_field()
This addresses a part of the code base that had a lot of type ignores.
This was due to how the return type of the `classify_field()` function
did not maintain the relationship of its two values, and not having
precise enough information for the case when the inner type is a subtype
of `Entity`.

By introducing specialized types for each four cases in conjunction with
a union return type, mypy is able to much better keep track of the
relationship that a field classified as an "entity field" must carry a
type that is a subtype of `Entity`.

As can be seen in the diff, test_introspect.py is the only changed test
module, this only has fallout on the internal introspection API. The
improvement and intent of this commit is to reduce the amount of type
ignores in `get_field_reader()` and `get_field_writer()`.