fix: Make UUID introspection independent of optionality
To accomodate for tagged UUID fields, where we have forced optionality
for two reasons (all tagged fields are optional, all UUID fields are
optional). We change the `get_reader()` implementation to not take
optionality into account for UUID fields.

This is in line with the corresponding `get_writer()` implementation.

A somewhat more semantically correct option to this would be to not
squash "is tag" and "is optional" into the same boolean value in
`get_field_reader`,  and rather keep track of that one level down.
However, that would cause the already large combinatorial realm of
possibilities to test for `get_reader()` to expand even larger. Living
the semantically slightly worse, but much less complex option seems like
the right thing to do.