feature: Build schema on upstream version 3.9.0-rc5 (#213)
This bumps schema to 3.9.0-rc5. The intention is that we can release
this as some pre-release, most likely an alpha, pending the upstream
final version.

This PR adds an `@xfail` to the compatibility test of
`UpdateRaftVoterResponse`. This model is hitting #215 which already
exists prior to schema vertsion 3.9.0. That bug should be addressed
separately.