fix: Correct the logic for tagged field defaults (#216)
This fixes an issue where the logic for tagged fields was not in sync
with upstream. This is in the realm of undocumented features of the
Kafka protocol, so there is no accurate documentation to reference. The
closest to documentation that exists, is [this section in a source code
README][messages-readme].


[messages-readme]:https://github.com/apache/kafka/tree/4419677e06b77ed9103c179d713450b1eb7881a1/clients/src/main/resources/common/message#deserializing-messages

There was a [long discussion] about this behavior on the Kafka dev
mailing list.

[long discussion]:
https://www.mail-archive.com/dev@kafka.apache.org/msg144437.html

This fixes #215.