feature: Add support for parsing record batches (#234)
Adds facilities for modeling, parsing, and serializing record batches
and records.

This will not add support for _automatically_ parsing and serializing
record batches off of e.g. FetchRequest and similar, and it's not clear
that that would even be a desired behavior. However, the introduced
facilities can be used in conjunction with the existing mechanisms and
will now provide a more complete set of building blocks allowing the
implementation of e.g. consumers and producers.